+++ Realm/RLMObjectStore.mm
@@ -143,6 +143,46 @@ void RLMRealmCreateAccessors(RLMSchema *schema) {
     }
 }
 
+NSError *RLMUpdateRealmToSchemaVersion(RLMRealm *realm, NSUInteger newVersion, RLMSchema *targetSchema, NSError *(^migrationBlock)()) {
+
+    [realm beginWriteTransaction];
+
+    @try {
+        NSUInteger oldVersion = RLMRealmSchemaVersion(realm);
+
+        // validate versions
+        if (oldVersion > newVersion && oldVersion != RLMNotVersioned) {
+            @throw [NSException exceptionWithName:@"RLMException"
+                                           reason:@"Version of Realm file on disk is higher than current schema version"
+                                         userInfo:@{@"path" : realm.path}];
+        }
+
+        // create tables
+        bool migrating = oldVersion != newVersion;
+        RLMRealmCreateTables(realm, targetSchema, migrating);
+
+        // apply migration block if provided
+        if (migrating && migrationBlock) {
+            NSError *error = migrationBlock();
+            if (error) {
+                [realm cancelWriteTransaction];
+                return error;
+            }
+        }
+
+        // set new version
+        RLMRealmSetSchemaVersion(realm, newVersion);
+    }
+    @catch (NSException *) {
+        [realm cancelWriteTransaction];
+        @throw;
+    }
+    
+    [realm commitWriteTransaction];
+    return nil;
+}
+
+
 void RLMRealmSetSchema(RLMRealm *realm, RLMSchema *targetSchema, bool verify) {
     realm.schema = [targetSchema copy];
 
@@ -159,6 +199,9 @@ void RLMRealmSetSchema(RLMRealm *realm, RLMSchema *targetSchema, bool verify) {
 
 
 void RLMRealmCreateTables(RLMRealm *realm, RLMSchema *targetSchema, bool updateExisting) {
+    // create metadata tables if neded
+    RLMRealmCreateMetadataTables(realm);
+
     realm.schema = [targetSchema copy];
 
     // first pass to create missing tables